+++ codlic.lisp
@@ -18,7 +18,6 @@
     ("skip-file-on-error" :none nil)
     ("skip-shebang" :none nil)
     ("print-license" :none nil)
-    ("license-search" :required nil)
     ("license-replace" :required nil)))
 
 (defun assoc-equal (item alist)
@@ -177,42 +176,20 @@ OPTIONS. May signal a `license-error'."
                    'license-error
                    :text "Failed to write to file."))))
 
-
-(defun replace-all (string part replacement &key (test #'char=))
-  "From the Common Lisp Cookbook, replaces instances of part with replacement."
-  (with-output-to-string (out)
-    (loop with part-length = (length part)
-       for old-pos = 0 then (+ pos part-length)
-       for pos = (search part string
-                         :start2 old-pos
-                         :test test)
-       do (write-string string out
-                        :start old-pos
-                        :end (or pos (length string)))
-       when pos do (write-string replacement out)
-       while pos)))
-
-(defun replace-lines (lines part replacement)
-  "Goes through each line in the array LINES and replaces all instances of PART
-with REPLACEMENT."
-  (map 'vector
-       (lambda (line)
-         (replace-all line part replacement))
-       lines))
-
 (defun read-license-lines (license-file opts)
   "Essentially calls READ-FILE-LINES to read the lines of LICENSE-FILE and
 return them. Also makes the necessary string replacements based on OPTS."
   (let ((license-lines (read-file-lines license-file))
-        (search-opt (assoc-equal "license-search" opts))
         (replace-opt (assoc-equal "license-replace" opts)))
-    (if search-opt
-        (replace-lines license-lines
-                       (cdr search-opt)
-                       (if replace-opt
-                           (cdr replace-opt)
-                           ""))
-        license-lines)))
+    (when replace-opt
+      (loop for replace-pair in (parse-search-replace-string (cdr replace-opt)) do
+           (when (string= (car replace-pair) "")
+             (license-error "Attempting to replace empty string."))
+           (setf license-lines
+                 (replace-lines license-lines
+                                (car replace-pair)
+                                (cdr replace-pair)))))
+    license-lines))
 
 (defun license-arg (arg options)
   "Licenses an argument based on the given options."