+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 This project adheres to Semantic Versioning
 
+## [Unreleased]
+### Fixed
+- Fix some errors in documentation.
+
+### Added
+- Add new search-replace feature for the license. This replaces the options
+  --license-search and --license-replace from before. Now it can do multiple
+  searches.
+
 ## [0.1.1] - 2017-06-09
 ### Fixed
 - Fix numerous errors in README, including incorrect executable name and